requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
chess==1.11.1
cssselect2==0.7.0
defusedxml==0.7.1
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
pygame==2.6.1
python-chess==1.999
tinycss2==1.4.0
twitchio==2.10.0
typing_extensions==4.12.2
webencodings==0.5.1
yarl==1.17.1
